Add support for tracing asynchronous code
This change adds AsyncioTracer, which is chosen instead of
SynchronousTracer as the tracer implementation if beeline is
initialized from within an asyncio event loop.

AsyncioTracer stores its trace state in a context variable. This
requires the contextvars module, which was added in Python 3.7.

AsyncioTracer also sets up the task factory for the event loop to hook
in a copying of the trace state for each new task. This ensures that
trace state behaves as expected for parallel running tasks.

The CI jobs have been updated to handle that some of the new code uses
async/await syntax that wasn't available in Python 2.7.